Imported Upstream version 1.3.0

+++ src/download/httpconnection.cpp
@@ -306,7 +306,7 @@ namespace bt
 		return true;
 	}
 	
-	float HttpConnection::getDownloadRate() const
+	int HttpConnection::getDownloadRate() const
 	{
 		if (sock)
 		{
